requirements.txt
Django
Pillow
django-crispy-forms
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
idna
oauthlib
Pillow
pycparser
PyJWT
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
urllib3
